feat: add LeadBodyComponent
