+++ packages/apps/crm/src/components/organisms/index.js
@@ -2,3 +2,4 @@ export {default as DropdownContactView} from './DropdownContactView/DropdownCont
 export {default as DropdownEventView} from './DropdownEventView/DropdownEventView';
 export {default as OpportunityDropdownInfo} from './OpportunityDropdownInfo/OpportunityDropdownInfo';
 export {default as LeadHeader} from './LeadHeader/LeadHeader';
+export {default as LeadBody} from './LeadBody/LeadBody';
